requirements.txt
python-decouple~=3.8
Django~=5.1.3
djangorestframework~=3.15.2